Add files via upload
I created an ideal gas law calculator using python and the tkinter GUI to build an simple UI that asks the user to select units and type in values.

The python script contains defined functions that takes user inputs and calculates the value from the user input and unit selection. Based on the unit selection, the gas constant r is checked in a dictionary to ensure unit combinations retrieve the correct value for r, and an error is shown to the user if the unit combination is incorrect. 

The functions also error check to ensure the user submits correct unit combinations. To ensure user inputs are valid, values are checked for type to ensure values are of type int or type float.

The three buttons on the program: 'Quit' to allow the program to be closed. 'Calculate' executes the function to calculate the value from user inputs. 'Clear' to delete all user inputs.

The python script is provided with documented code that describes how each function works, and the .exe is provided to show how the code works with the tkinter GUI